program to check for recurring string in array of strings
Alice is asked to find the total number of occurences of a particular
string given an array of strings. Each string is made of characters
ranging from A to Z.

Two strings s1 and s2 are considered same if s1 can be converted to s2
in one move or s2 can be converted to s1 in one move

A move is defined as follows:
Move: replacing any one character of a string with the character that
occurs immediately next to it in the english alphabet. e.g. A->B, B->C,
... Z->A. "->" represents "replaced with"

Input specs:

input1: N, denoting the total numbers of strings in the array
input2: An array of N strings
input3: an string, s1, whose total occurences are to be determined.

Ouput specs:
Your function should return the number of occurences of s1 in the given
array of strings